requirements-uv.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
datasets==4.1.1
dill==0.4.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
gdown==5.2.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
nltk==3.9.2
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.3.2
psutil==7.1.0
pyarrow==21.0.0
pyparsing==3.2.5
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
rouge-score==0.1.2
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
triton==3.4.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.20.1
